add 100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int i = 1;
-	unsigned long long prime = 612852475143;
+	unsigned long prime = 612852475143;
 	int m = 3;
 
 	while (i > 0)
